Fixed bug in test

+++ contracts/NFTMarketplace.sol
@@ -15,7 +15,7 @@ contract NFTMarketplace is ERC721URIStorage {
    Counters.Counter private _tokenIds;
    Counters.Counter private _itemsSold;
 
-   uint listingPrice = 0.025 ether;
+   uint256 listingPrice = 0.025 ether;
    address payable owner;
 
    mapping(uint256 => MarketItem) private idToMarketItem;
@@ -124,7 +124,7 @@ contract NFTMarketplace is ERC721URIStorage {
       uint currentIndex = 0;
 
       MarketItem[] memory items = new MarketItem[](unsoldItemCount);
-      for (uint i =0; i < itemCount; i++) {
+      for (uint i = 0; i < itemCount; i++) {
          if (idToMarketItem[i + 1].owner == address(this)) {
                uint currentId = i + 1;
                MarketItem storage currentItem = idToMarketItem[currentId];
@@ -149,7 +149,7 @@ contract NFTMarketplace is ERC721URIStorage {
 
       MarketItem[] memory items = new MarketItem[](itemCount);
       for (uint i = 0; i < totalItemCount; i++) {
-         if (idToMarketItem[i + 1]. owner == msg.sender) {
+         if (idToMarketItem[i + 1].owner == msg.sender) {
             uint currentId = i + 1;
             MarketItem storage currentItem = idToMarketItem[currentId];
             items[currentIndex] = currentItem;
